fix: vue is not available for extensions in voila-vuetify
Export vue from the nodeps-bundle just like the regular bundle, so
vue is available for dependent extensions in voila-vuetify. It
might seem unnecessary because vue is not included in this bundle,
but dependent extensions still try to load it from here. This is
due to the webpack alias used in dependent extensions.

The error did not occur for the nodeps version of jupyter-vuetify,
because it only uses the vueRender() function of the nodeps-bundle.