fix: self-closing tags don't work properly
Since 66011a33 Vue templates are parsed by the browser. This means
the template is handled as HTML, which doesn't support self-closing
tags for custom elements and results in siblings after the illegal
self-closing tag not being rendered correctly.

This commit uses the single file component parser of the Vue
compiler to fix this.