fix: support for traitlets 5.0.5
Setting the default of a traitlets.List to None resulted in the
actual defalt of an empty list. This changed trailtets 5.0.5 where
the it would actually be None. This caused a problem for _events
which should be [] instead of the specified None.

The other lines are changed so they behave the same as before
5.0.5.

See #18