fix: render ipywidgets property as a vue child component
There were two things that could go wrong when rendering an ipywidget
as a child of the vue component using the createObjectForNestedModel.
First, since create_child_view is async, the vue component could
be destroyed before the child view is created. In that case,
we should not attach the view, but directory remove it again.

Second, child views were not removed properly, since we did not listen
to the destroyed/beforeDestroy life cycled events of the vue component.

We now remove the ipywidget view when the vue component is destroyed,
but only when it's already created.